requirements.txt
Django==1.7.2
Pillow==2.7.0
django-model-utils==2.2
django-photologue==3.1.1
django-sortedm2m==0.8.1
pytz==2014.10
